用户处理

+++ src/main/java/com/lange/game/domian/Plate.java
@@ -4,6 +4,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,7 +15,7 @@ public class Plate {
     private Long id;
     private Long gameId;//比赛id
     private String content;//盘口内容
-    private Double odds;//赔率
+    private BigDecimal odds;//赔率
     private Integer type;//0:可下注 1:盘口过期 3：已赔付
     private Date createTime;//
     private Date endTime;//
